requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
# langchain-chroma==0.2.2
# langchain-pinecone==0.2.0
# langchain-text-splitters==0.3.7
# langgraph==0.3.21
# langsmith==0.3.19
openai==1.69.0
python-dotenv==1.1.0
line-bot-sdk==3.17.1
Flask[async]==3.0.3